Revert "Оптимизированная работа в зависимости от настройки использования БД Database.ConfigurationIsEnabled"
This reverts commit c1e6fa459efb759dba8b73a1a99f6280b8ed5a6c.

+++ src/Core/FmuApiApplication/Documents/CancelDocument.cs
@@ -73,6 +73,9 @@ namespace FmuApiApplication.Documents
             if (!_configuration.Database.ConfigurationIsEnabled)
                 return Result.Success(checkResult);
 
+            if (_configuration.Database.FrontolDocumentsDbName == string.Empty)
+                return Result.Success(checkResult);
+
             await _markInformationService.DeleteDocumentFromDbAsync(_document.Uid);
 
             return Result.Success(checkResult);
